requirements.txt
streamlit==1.40.0
anthropic==0.39.0
httpx==0.27.0  # Pin to compatible version
requests==2.32.3
pandas==2.2.3
plotly==5.24.1
python-dotenv==1.0.1
pytest==8.3.4
reportlab==4.2.5
beautifulsoup4==4.12.3
lxml==5.3.0
tenacity==9.0.0
pydantic==2.10.4